Split codebase into packages
