+++ go.mod
@@ -3,7 +3,6 @@ module Pelmenner/TransferBot
 go 1.18
 
 require (
-	github.com/SevereCloud/vksdk v1.10.0
 	github.com/SevereCloud/vksdk/v2 v2.14.0
 	github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1
 	github.com/mattn/go-sqlite3 v1.14.13
@@ -11,7 +10,6 @@ require (
 
 require (
 	github.com/klauspost/compress v1.15.6 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/vmihailenco/msgpack/v5 v5.3.5 // indirect
 	github.com/vmihailenco/tagparser/v2 v2.0.0 // indirect
 	golang.org/x/text v0.3.7 // indirect
